Pre-commit: fix files
