+++ tests/nodes/connectors/conftest.py
@@ -13,7 +13,7 @@ import docker
 from docker.models.containers import Container
 import pytest
 
-from .opcua_test_server import create_server
+from . import create_server
 
 
 @pytest.fixture(scope="session")
